requirements.txt
# Dependencies for VMR Dashboard
streamlit>=1.30.0
gspread>=5.12.0
google-auth>=2.23.0
plotly>=5.18.0
pandas>=2.0.0
